// packages/pn-personagiuridica-webapp/src/models/PGEventsType.ts
export enum PGEventsType {
  /* API KEYS */
  SEND_PG_API_INTEGRATION = 'SEND_PG_API_INTEGRATION',
  SEND_PG_ADD_API_START = 'SEND_PG_ADD_API_START',
  SEND_PG_ADD_API_UX_SUCCESS = 'SEND_PG_ADD_API_UX_SUCCESS',

  /* CONTACT */
  SEND_PG_YOUR_CONTACT_DETAILS = 'SEND_PG_YOUR_CONTACT_DETAILS',
  // SERCQ
  SEND_PG_ADD_SERCQ_START = 'SEND_PG_ADD_SERCQ_START',
  SEND_PG_ADD_SERCQ_UX_SUCCESS = 'SEND_PG_ADD_SERCQ_UX_SUCCESS',
  SEND_PG_REMOVE_SERCQ_START = 'SEND_PG_REMOVE_SERCQ_START',
  SEND_PG_REMOVE_SERCQ_UX_SUCCESS = 'SEND_PG_REMOVE_SERCQ_UX_SUCCESS',
  // EMAIL
  SEND_PG_ADD_EMAIL_START = 'SEND_PG_ADD_EMAIL_START',
  SEND_PG_ADD_EMAIL_UX_SUCCESS = 'SEND_PG_ADD_EMAIL_UX_SUCCESS',
  SEND_PG_REMOVE_EMAIL_START = 'SEND_PG_REMOVE_EMAIL_START',
  SEND_PG_REMOVE_EMAIL_UX_SUCCESS = 'SEND_PG_REMOVE_EMAIL_UX_SUCCESS',
  // SMS
  SEND_PG_ADD_SMS_START = 'SEND_PG_ADD_SMS_START',
  SEND_PG_ADD_SMS_UX_SUCCESS = 'SEND_PG_ADD_SMS_UX_SUCCESS',
  SEND_PG_REMOVE_SMS_START = 'SEND_PG_REMOVE_SMS_START',
  SEND_PG_REMOVE_SMS_UX_SUCCESS = 'SEND_PG_REMOVE_SMS_UX_SUCCESS',

  /* MANDATE */
  SEND_PG_ADD_MANDATE_START = 'SEND_PG_ADD_MANDATE_START',
  SEND_PG_ADD_MANDATE_UX_SUCCESS = 'SEND_PG_ADD_MANDATE_UX_SUCCESS',
  SEND_PG_MANDATES_RECEIVED = 'SEND_PG_MANDATES_RECEIVED',
  SEND_PG_MANDATES_GIVEN = 'SEND_PG_MANDATES_GIVEN',

  /* NAVIGATION */
  SEND_PG_OPEN_USERS = 'SEND_PG_OPEN_USERS',
  SEND_PG_OPEN_GROUPS = 'SEND_PG_OPEN_GROUPS',
  SEND_PG_HELP = 'SEND_PG_HELP',
  SEND_PG_EXIT = 'SEND_PG_EXIT',

  /* NOTIFICATION */
  SEND_PG_YOUR_NOTIFICATION = 'SEND_PG_YOUR_NOTIFICATION',
  SEND_PG_NOTIFICATION_DETAIL = 'SEND_PG_NOTIFICATION_DETAIL',
  SEND_PG_NOTIFICATION_DELEGATED = 'SEND_PG_NOTIFICATION_DELEGATED',

  /* SERVICE STATUS */
  SEND_PG_SERVICE_STATUS = 'SEND_PG_SERVICE_STATUS',

  /* SUPER PROPERTIES */
  USER_ROLE = 'USER_ROLE',
  SEND_PG_HAS_EMAIL = 'SEND_PG_HAS_EMAIL',
  SEND_PG_HAS_SMS = 'SEND_PG_HAS_SMS',
  SEND_PG_HAS_SERCQ = 'SEND_PG_HAS_SERCQ',
  SEND_PG_HAS_MANDATE = 'SEND_PG_HAS_MANDATE',
  SEND_PG_HAS_MANDATE_GIVEN = 'SEND_PG_HAS_MANDATE_GIVEN',
  SEND_PG_HAS_NOTIFICATIONS = 'SEND_PG_HAS_NOTIFICATIONS',
}